+++ dummy/src/main.rs
@@ -1,35 +1,23 @@
 use nutype::nutype;
 
-fn validate_name(name: &str) -> Result<(), NameError> {
-    if name.len() < 3 {
-        Err(NameError::TooShort)
-    } else if name.len() > 10 {
-        Err(NameError::TooLong)
-    } else {
-        Ok(())
-    }
-}
-
-#[derive(Debug, PartialEq)]
-enum NameError {
-    TooShort,
-    TooLong,
-}
-
 #[nutype(
-    validate(with = validate_name, error = NameError),
-    derive(Debug, AsRef, PartialEq),
+    const_fn,
+    derive(Debug),
+    validate(greater_or_equal = -273.15),
 )]
-struct Name(String);
+pub struct Celsius(f64);
+
+macro_rules! nutype_const {
+    ($name:ident, $ty:ty, $value:expr) => {
+        const $name: $ty = match <$ty>::try_new($value) {
+            Ok(value) => value,
+            Err(_) => panic!("Invalid value"),
+        };
+    };
+}
+
+nutype_const!(WATER_BOILING_POINT, Celsius, 100.0);
 
 fn main() {
-    let name = Name::try_new("John").unwrap();
-    assert_eq!(name.as_ref(), "John");
-
-    assert_eq!(
-        Name::try_new("JohnJohnJohnJohnJohn"),
-        Err(NameError::TooLong)
-    );
-
-    assert_eq!(Name::try_new("Jo"), Err(NameError::TooShort));
+    println!("{:?}", WATER_BOILING_POINT);
 }