+++ nutype_macros/src/integer/gen/mod.rs
@@ -24,7 +24,7 @@ use crate::common::{
         traits::GeneratedTraits,
         GenerateNewtype,
     },
-    models::{ErrorTypePath, Guard, TypeName},
+    models::{ConstFn, ErrorTypePath, Guard, TypeName},
 };
 
 impl<T> GenerateNewtype for IntegerNewtype<T>
@@ -39,6 +39,7 @@ where
     fn gen_fn_sanitize(
         inner_type: &Self::InnerType,
         sanitizers: &[Self::Sanitizer],
+        const_fn: ConstFn,
     ) -> TokenStream {
         let transformations: TokenStream = sanitizers
             .iter()
@@ -55,7 +56,7 @@ where
             .collect();
 
         quote!(
-            fn __sanitize__(mut value: #inner_type) -> #inner_type {
+            #const_fn fn __sanitize__(mut value: #inner_type) -> #inner_type {
                 #transformations
                 value
             }
@@ -66,6 +67,7 @@ where
         inner_type: &Self::InnerType,
         error_type_path: &ErrorTypePath,
         validators: &[Self::Validator],
+        const_fn: ConstFn,
     ) -> TokenStream {
         let validations: TokenStream = validators
             .iter()
@@ -109,7 +111,7 @@ where
             .collect();
 
         quote!(
-            fn __validate__(val: &#inner_type) -> ::core::result::Result<(), #error_type_path> {
+            #const_fn fn __validate__(val: &#inner_type) -> ::core::result::Result<(), #error_type_path> {
                 let val = *val;
                 #validations
                 Ok(())