+++ nutype_macros/src/common/gen/mod.rs
@@ -316,12 +316,6 @@ pub trait GenerateNewtype {
                 // scope imported with `use super::*`.
                 #fn_sanitize
                 #fn_validate
-
-                // TODO: Remove in 0.5.0
-                #[deprecated(since="0.4.3", note="\nUse `try_new` instead.")]
-                pub fn new(raw_value: #input_type) -> ::core::result::Result<Self, #error_type_path> {
-                    Self::try_new(raw_value)
-                }
             }
         )
     }