+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.x.x - 2025-xx-xx
-* **[FEATURE]** Ability to instantiate types in `const` context, when declared with `const_fn` flag.
+- **[FEATURE]** Ability to instantiate types in `const` context, when declared with `const_fn` flag.
+- **[BREAKING]** Fallible `::new()` constructor is removed (was deprecated in 0.4.3).
 
 ### v0.5.1 - 2024-12-20