+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-### v0.x.x - 2025-xx-xx
-- **[BREAKING]** Fallible `::new()` constructor is removed (was deprecated in 0.4.3).
+### v0.6.0 - 2025-02-02
 - **[FEATURE]** Ability to instantiate types in `const` context, when declared with `const_fn` flag.
 - **[FEATURE]** Support `derive(IntoIterator)` for inner types that implement `IntoIterator`.
 - **[FIX]** Enable `&'a str` as an inner type.
+- **[BREAKING]** Fallible `::new()` constructor is removed (was deprecated in 0.4.3).
 
 ### v0.5.1 - 2024-12-20